Fixed feedback_key problem

+++ ws/libresmack_fuzz/include/resmack/fuzz/corpus.hpp
@@ -8,7 +8,7 @@ namespace fuzz {
 
 class Corpus {
  public:
-  virtual void AddRandSnapshot(resmack::Vector<RandSnapshot>* snapshot, uint32_t feedback_key) = 0;
+  virtual void AddRandSnapshot(resmack::Vector<RandSnapshot>* snapshot, size_t feedback_key) = 0;
   virtual Vector<RandSnapshot>* GetItem(Rand* rand) = 0;
   /**
    * Intended to be called on intervals to do "processing" (whatever that means
@@ -16,10 +16,10 @@ class Corpus {
    **/
   virtual void Sync() = 0;
   virtual size_t NumItems() = 0;
-  virtual bool SeenFeedback(uint32_t key) = 0;
+  virtual bool SeenFeedback(size_t key) = 0;
   virtual bool AddRandSnapshotIfNotSeen(
     resmack::Vector<RandSnapshot>* snapshot,
-    uint32_t key
+    size_t key
   ) = 0;
 };
